Moved most of strelay code to lib.
